# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ID8 Operator API
  version: 0.1.0
  description: |
    MVP API contract for orchestrating prompt-to-production runs with HITL gates.
servers:
  - url: https://api.id8.local
paths:
  /v1/projects:
    get:
      summary: List projects with latest run summaries
      operationId: listProjects
      responses:
        '200':
          description: Project list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
    post:
      summary: Create a project from a user prompt
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /v1/projects/{projectId}:
    get:
      summary: Get project status and metadata
      operationId: getProject
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Project found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
...
  /v1/projects/{projectId}/design/generate:
    post:
      summary: Generate or regenerate design artifacts
      operationId: generateDesign
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DesignGenerateRequest'
      responses:
        '202':
          description: Design job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactResponse'
...
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
  schemas:
    DesignProvider:
      type: string
      enum:
        - stitch_mcp
        - internal_spec
        - manual_upload
    ModelProfile:
      type: string
      enum:
        - primary
        - customtools
        - fallback
    ProjectStatus:
      type: string
      enum:
        - ideation
        - prd_draft
        - prd_approved
        - design_draft
        - design_approved
        - tech_plan_draft
        - tech_plan_approved
        - codegen
        - security_gate
        - deploy_ready
        - deploying
        - deployed
        - failed
    ArtifactType:
      type: string
      enum:
        - prd
        - design_spec
        - tech_plan
        - code_snapshot
        - security_report
        - deploy_report
...
    StitchAuthMethod:
      type: string
      enum:
        - api_key
        - oauth_access_token
    StitchAuthPayload:
      type: object
      properties:
        auth_method:
          $ref: '#/components/schemas/StitchAuthMethod'
        api_key:
          type: string
        oauth_token:
          type: string
        goog_user_project:
          type: string
...
    CreateProjectRequest:
      type: object
      properties:
        initial_prompt:
          type: string
        constraints:
          type: object
          additionalProperties: true
      required:
        - initial_prompt
...
    DesignGenerateRequest:
      type: object
      properties:
        provider:
          $ref: '#/components/schemas/DesignProvider'
        model_profile:
          $ref: '#/components/schemas/ModelProfile'
        prompt_constraints:
          type: object
          additionalProperties: true
        stitch_auth:
          $ref: '#/components/schemas/StitchAuthPayload'
      required:
        - provider
        - model_profile
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_user_id:
          type: string
          format: uuid
        initial_prompt:
          type: string
        status:
          $ref: '#/components/schemas/ProjectStatus'
        github_repo_url:
          type: string
        live_deployment_url:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - owner_user_id
        - initial_prompt
        - status
        - created_at
        - updated_at
    ProjectRunSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/ProjectStatus'
        current_node:
          type: string
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - status
        - current_node
        - updated_at
    ProjectListItem:
      allOf:
        - $ref: '#/components/schemas/Project'
        - type: object
          properties:
            latest_run:
              $ref: '#/components/schemas/ProjectRunSummary'
    ProjectListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProjectListItem'
      required:
        - items
...
    ProjectArtifact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        run_id:
          type: string
          format: uuid
        artifact_type:
          $ref: '#/components/schemas/ArtifactType'
        version:
          type: integer
        content:
          type: object
          additionalProperties: true
        model_profile:
          $ref: '#/components/schemas/ModelProfile'
        created_at:
          type: string
          format: date-time
      required:
        - id
        - project_id
        - run_id
        - artifact_type
        - version
        - content
        - created_at
    ArtifactResponse:
      type: object
      properties:
        artifact:
          $ref: '#/components/schemas/ProjectArtifact'
      required:
        - artifact
...